+++ src/EtheriT.Coker.Application.Shared/Dto/Order/OrderHeaderDataDto.cs
@@ -6,13 +6,13 @@ namespace EtheriT.Coker.Application.Shared.Dto.Order
     public class OrderHeaderDataDto
     {
         public long Id { get; set; }
-        public int Subtotal { get; set; }
-        public int? Discount { get; set; }
+        public decimal Subtotal { get; set; }
+        public decimal? Discount { get; set; }
         public int? Bonus { get; set; }
         public long? CouponId { get; set; }
         public int? Service_Charge { get; set; }
-        public int Freight { get; set; }
-        public int? Total { get; set; }
+        public decimal Freight { get; set; }
+        public decimal? Total { get; set; }
         public OrderStatusEnum State { get; set; }
         public string Orderer { get; set; }
         public SexEnum? OrdererSex { get; set; }
